v1/requirements_extractor.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
PyMuPDF>=1.18.0
pdfplumber>=0.5.28
Pillow>=8.0.0
pytesseract>=0.3.7